+++ src/main.rs
@@ -3,5 +3,5 @@ use mylang::io::process_files;
 use std::{env, process};
 
 fn main() -> Result<()> {
-    process::exit(process_files(env::args().skip(1).collect())? as i32)
+    process::exit(process_files(env::args().skip(1).collect())?)
 }
